requirements.txt
flask
google-generativeai
requests
gunicorn


